moved cmd_util.py to apio/utils
