+++ apio/commands/apio_drivers_ftdi.py
@@ -11,7 +11,7 @@ import sys
 import click
 from apio.managers.drivers import Drivers
 from apio.apio_context import ApioContext, ApioContextScope
-from apio.cmd_util import ApioGroup, ApioSubgroup
+from apio.utils.cmd_util import ApioGroup, ApioSubgroup
 from apio.managers.system import System
