+++ apio/commands/apio_graph.py
@@ -11,7 +11,7 @@ import sys
 from pathlib import Path
 import click
 from apio.managers.scons import SCons
-from apio import cmd_util
+from apio.utils import cmd_util
 from apio.commands import options
 from apio.apio_context import ApioContext, ApioContextScope
 from apio.utils.util import nameof
