+++ apio/commands/apio_lint.py
@@ -10,7 +10,7 @@ import sys
 from pathlib import Path
 import click
 from apio.managers.scons import SCons
-from apio import cmd_util
+from apio.utils import cmd_util
 from apio.commands import options
 from apio.apio_context import ApioContext, ApioContextScope
