+++ apio/commands/apio_packages.py
@@ -14,7 +14,7 @@ from apio.managers import installer
 from apio.apio_context import ApioContext, ApioContextScope
 from apio import pkg_util
 from apio.commands import options
-from apio.cmd_util import ApioGroup, ApioSubgroup
+from apio.utils.cmd_util import ApioGroup, ApioSubgroup
 
 
 # ------ apio packages install
