+++ apio/commands/apio_system.py
@@ -11,7 +11,7 @@ import click
 from click import secho
 from apio.utils import util
 from apio.apio_context import ApioContext, ApioContextScope
-from apio.cmd_util import ApioGroup, ApioSubgroup
+from apio.utils.cmd_util import ApioGroup, ApioSubgroup
 
 
 # ------ apio system info
