+++ apio/commands/apio_upload.py
@@ -12,7 +12,7 @@ from pathlib import Path
 import click
 from apio.managers.scons import SCons
 from apio.managers.drivers import Drivers
-from apio import cmd_util
+from apio.utils import cmd_util
 from apio.commands import options
 from apio.apio_context import ApioContext, ApioContextScope
